+++ src/main/java/ddd/caffeine/ratrip/module/auth/application/impl/KakaoAuthService.java
@@ -12,8 +12,8 @@ import ddd.caffeine.ratrip.module.auth.application.dto.SignInDto;
 import ddd.caffeine.ratrip.module.auth.application.dto.SignUpDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.response.SignInResponseDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.response.TokenResponseDto;
-import ddd.caffeine.ratrip.module.feign.auth.kakao.KakaoUserApiClient;
-import ddd.caffeine.ratrip.module.feign.auth.kakao.dto.KakaoProfileResponse;
+import ddd.caffeine.ratrip.module.feign.auth.kakao.user.KakaoUserApiClient;
+import ddd.caffeine.ratrip.module.feign.auth.kakao.user.dto.KakaoProfileResponse;
 import ddd.caffeine.ratrip.module.user.application.UserService;
 import ddd.caffeine.ratrip.module.user.application.dto.RegisterUserDto;
 import ddd.caffeine.ratrip.module.user.application.dto.SignInUserDto;
