--- src/main/java/ddd/caffeine/ratrip/module/feign/auth/kakao/dto/KakaoAccount.java
@@ -1,9 +1,0 @@
-package ddd.caffeine.ratrip.module.feign.auth.kakao.dto;
-
-import lombok.Getter;
-
-@Getter
-public class KakaoAccount {
-	private Profile profile;
-	private String email;
-}
