--- src/main/java/ddd/caffeine/ratrip/module/feign/auth/kakao/dto/Profile.java
@@ -1,9 +1,0 @@
-package ddd.caffeine.ratrip.module.feign.auth.kakao.dto;
-
-import lombok.Getter;
-
-@Getter
-public class Profile {
-	private String nickname;
-}
-
